+++ emborg/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.39"
-__released__ = "2024-04-29"
+__version__ = "1.39.1"
+__released__ = "2024-05-27"
 
 from .emborg import Emborg